Seperated behaviours between watching and monitoring.
