+++ package.json
@@ -14,6 +14,7 @@
     "jade": "~1.11.0",
     "morgan": "~1.7.0",
     "nedb": "^1.8.0",
-    "serve-favicon": "~2.3.0"
+    "serve-favicon": "~2.3.0",
+    "winston": "^2.2.0"
   }
 }
