+++ routes/api.js
@@ -1,11 +1,16 @@
 var express = require('express');
 var router = express.Router();
 var metrics = require('../controllers/metrics');
+var db = require('../controllers/datastores');
+var watch = require('../controllers/watch')();
 
-metrics.setCheckRate(0.5);
+watch.setCheckRate(5);
+
+// metrics.setCheckRate(0.5);
 /* GET users listing. */
 router.get('/scan', function(req, res, next) {
   metrics.networkScan(function (data) {
+    db.devices.updateDevices(data, callback);
     res.send(data);
   });
 });
